third_party/blink/web_tests/platform/mac/external/wpt/css/selectors/parsing/parse-state-expected.txt

This is a testharness.js-based test.
[FAIL] ":state(--foo)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': ':state(--foo)' is not a valid selector.
[FAIL] ":state(bar)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': ':state(bar)' is not a valid selector.
[FAIL] ":state(--)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': ':state(--)' is not a valid selector.
[FAIL] ":state(--0)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': ':state(--0)' is not a valid selector.
[FAIL] ":host(:state(--foo))" should be a valid selector
  Failed to execute 'querySelector' on 'Document': ':host(:state(--foo))' is not a valid selector.
[FAIL] "my-input[type=\\"foo\\"]:state(checked)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': 'my-input[type="foo"]:state(checked)' is not a valid selector.
[FAIL] "my-input[type=\\"foo\\"]:state(--0)::before" should be a valid selector
  Failed to execute 'querySelector' on 'Document': 'my-input[type="foo"]:state(--0)::before' is not a valid selector.
[FAIL] "my-input[type=\\"foo\\"]:state(--0)::part(inner)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': 'my-input[type="foo"]:state(--0)::part(inner)' is not a valid selector.
[FAIL] "my-input[type=\\"foo\\"]:state(--0)::part(inner):state(bar)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': 'my-input[type="foo"]:state(--0)::part(inner):state(bar)' is not a valid selector.
[FAIL] "::part(inner):state(bar)::before" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::part(inner):state(bar)::before' is not a valid selector.
[FAIL] "::part(inner):state(bar)::after" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::part(inner):state(bar)::after' is not a valid selector.
Harness: the test ran to completion.
